refactor : DTO 객체를 dto 패키지로 이동
- 나머지 자잘한 변경들을 함께 반영한다
